Fix issue for messages of more than 512 MB
Before this fix the maximum message length was 0x1FFFFFFF bytes.
Starting with length of 0x20000000 bytes (i.e., 512 MB) and greater the
result was no longer valid.

A new test vector has been added for the hash of a message of 1GB. Add
-DTEST_VECTORS_LONG to the Makefile to enable the test of this vector.

There is still a limit of max 4 GB as 32-bit types are used. This will
be fixed in a new commit with the use of 64-bit types.

BUG: #1, #3