Fix regression issue
A regression was introduced in 82a5a95 when adding support for 64-bit
type for length.

We removed the line that caused the issue. In some case the object
assignment could overwrite algorithm internal data. The line was
supposed to handle messages larger than 2^61 bytes, but such sizes are
beyond processing capabilities anyway.

We also added a new test vector for this issue #4.

BUG: #4

54aa4320c4a2e56492871ad5df3e600507